+++ apps/desktop/renderer/src/design-system/templates/index.ts
@@ -1,0 +1,3 @@
+export * from './WindowChrome';
+export * from './AppShell';
+export * from './SplitMarketingForm';